refactor(svg): #67 - ThreeDotsIcon를 공통 폴더로 이동
